// src/lib/models/types.ts
export enum TypeComponent {
	CARD = "CARD",
	TABLE = "TABLE",
	TEXT = "TEXT",
	BUTTON = "BUTTON",
	COLUMNS = "COLUMNS",
	LIST = "LIST",
	KEY_VALUE = "KEY_VALUE",
	DIVIDER = "DIVIDER",
	FLEX_CONTAINER = "FLEX_CONTAINER",
	GRID_LAYOUT = "GRID_LAYOUT",
	PANEL = "PANEL",
	CONTENT_BOX = "CONTENT_BOX",
	SLIDER = "SLIDER",
	SELECT = "SELECT",
	SWITCH = "SWITCH",
	SEND_TEXT = "SEND_TEXT",
	JSON = "JSON"
}